+++ test/geoao_test.dart
@@ -3,10 +3,13 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:geoao/geoao.dart';
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
+  test('test provinces and counties assets', () async {
+    TestWidgetsFlutterBinding.ensureInitialized();
+
+    final geoao = Geoao();
+    expect((await geoao.getCounties(1))[0].countyName, 'Belas');
+    // expect(calculator.addOne(2), 3);
+    // expect(calculator.addOne(-7), -6);
+    // expect(calculator.addOne(0), 1);
   });
 }